cloud/storage/dropbox: return errors.NotExist when block is not found
By introducing a custom DropboxAPIError type for handling API specific
errors in case of a 409 status code. The implementation is very basic
and looks for a "not_found" string inside the error message.

The reason is that the structure of the error JSON is dependent on the
API endpoints and the Upspin storage backends only return a 'errors.NotExist',
in case a block is not found, otherwise always with a 'errors.IO'.

Fix upspin/upspin#516

Change-Id: Iff4736fe0f8a406614a426926c1b8fe44c2c977d
Reviewed-on: https://upspin-review.googlesource.com/15382
Reviewed-by: Andrew Gerrand <adg@golang.org>